improved folder structure
