--- controllers/contacts.js
@@ -1,66 +1,0 @@
-const { Contact } = require("../models/contact");
-const HttpError = require("../helpers/HttpError");
-const ctrlWrapper = require("../helpers/ctrlWrapper");
-
-const getAll = async (req, res, next) => {
-  const result = await Contact.find();
-  res.json(result);
-};
-
-const getById = async (req, res, next) => {
-  const { contactId } = req.params;
-  const result = await Contact.findById(contactId);
-  if (!result) {
-    throw HttpError(404, "Not faund");
-  }
-  res.json(result);
-};
-
-const post = async (req, res, next) => {
-  const result = await Contact.create(req.body);
-  res.status(201).json(result);
-};
-
-const deleteById = async (req, res, next) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndRemove(contactId);
-  if (!result) {
-    throw HttpError(404, "Not faund");
-  } else {
-    res.status(200).json({
-      message: "Delete success",
-    });
-  }
-};
-
-const updateById = async (req, res, next) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.json(result);
-};
-
-const updateStatusContact = async (req, res) => {
-  const { contactId } = req.params;
-
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, "Not Found");
-  }
-  res.json(result);
-};
-
-module.exports = {
-  getAll: ctrlWrapper(getAll),
-  getById: ctrlWrapper(getById),
-  post: ctrlWrapper(post),
-  updateById: ctrlWrapper(updateById),
-  updateStatusContact: ctrlWrapper(updateStatusContact),
-  deleteById: ctrlWrapper(deleteById),
-};
